schleiden-virtual-robot/requirements.txt
PyQt5>=5.15
opencv-python
numpy
pyyaml
pyttsx3
SpeechRecognition
Pillow
textblob
